+++ backend/Routes/orders.js
@@ -2,7 +2,7 @@ const express = require("express")
 const router = express.Router();
 const orderModel = require("../Models/orderModel");
 const { auth,isAdmin,isUser } = require("../Middleware/auth");
-const {addOrder, updateOrder, getAllOrders, getByID, deleteById} = require("../Controllers/ordersControl");
+const {addOrder, updateOrder, getAllOrders, getByID, deleteById, updateOrderByAdmin} = require("../Controllers/ordersControl");
 
 
 
@@ -11,6 +11,8 @@ router.use(auth)
 router.post("/",isUser, addOrder);
 // update order 
 router.patch("/update/:id", isUser, updateOrder);
+// update order 
+router.patch("/updateAdmin/:id", isAdmin, updateOrderByAdmin);
 // get by id
 router.get("/:id", getByID);
 // get all orders
